+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 
 * `BadBeans` is now `RecipeError`. It has now an `Exception` instance and a pretty function.
 
+* `exportToDot` is now `writeAsDot` and accepts a `RecipeError` to highlight problematic nodes.
+
 * Now `Constructor`s and `Recipe`s keep track of the `CallStack` of when they were created. This is used
   by errors to print the relevant code locations.
   Because now we have code locations, `PathToCauldron` is no longer useful and has been removed.